+++ examples/parallel.js
@@ -1,4 +1,4 @@
-var scraper = require('scraper');
+var scraper = require('../lib/scraper');
 
 scraper([
 		'https://example.test/search?q=javascript'
@@ -14,9 +14,7 @@ scraper([
 	, function(err, $) {
 	if (err) {throw err;}
 
-	$('.msg').each(function() {
-		console.log($(this).text().trim()+'\n');
-	});
+	console.log($('title').text().trim());
 }, {
 	'reqPerSec': 0.2 // Wait 5sec between each external request
 });