+++ examples/simple.js
@@ -1,9 +1,7 @@
-var scraper = require('scraper');
+var scraper = require('../lib/scraper');
 
 scraper('https://example.test/search?q=javascript', function(err, $) {
 	if (err) {throw err;}
 
-	$('.msg').each(function() {
-		console.log($(this).text().trim()+'\n');
-	});
+	console.log($('title').text().trim());
 });