+++ examples/test.js
@@ -1,33 +1,39 @@
 var http = require('http');
-http.createServer(function (request, response) {
+var server = http.createServer(function (request, response) {
 	var respondTime = Math.floor(Math.random()*1000);
 	setTimeout(function() {
 		var links = '';
-		var linkCount = Math.floor(Math.random()*400)+100;
+		var linkCount = Math.floor(Math.random()*20)+5;
 		for (var i=0; i < linkCount; i++) {
 			links += '<a href="/">Test - '+i+'</a>';
-		};
-		response.writeHead(200, {'Content-Type': 'text/plain'});
+		}
+		response.writeHead(200, {'Content-Type': 'text/html'});
 		response.end('<html><head><title>Page has '+linkCount+' links</title></head><body><div id="time">'+respondTime+'</div>'+links+'</body></html>\n');
 	}, respondTime);
 }).listen(8486);
 
 var urls = [];
-for (var u=0; u < 2; u++) {
-	for (var i=0; i < 1000; i++) {
-		urls.push('http://localhost:8486');
-	};
-};
+var requestCount = parseInt(process.env.SCRAPER_TEST_REQUESTS || '4', 10);
+for (var u=0; u < requestCount; u++) {
+	urls.push('http://localhost:8486');
+}
 
 setTimeout(function() {
-	var scraper = require('scraper');
+	var scraper = require('../lib/scraper');
+	var remaining = urls.length;
 	scraper(urls, function(err, $) {
 		if (err) {
 			console.log(err);
+			if (--remaining === 0) {
+				server.close();
+			}
 			return;
 		}
 		console.log('Server took '+$('#time').text()+'ms to respond, found '+$('a').length+' <a> @ '+new Date().toString()+' with title "'+$('title').text()+'"');
+		if (--remaining === 0) {
+			server.close();
+		}
 	}, {
-		'reqPerSec': 0.5
+		'reqPerSec': parseFloat(process.env.SCRAPER_TEST_REQ_PER_SEC || '2')
 	});
 }, 1000);